// Website Files/.backups/backup_2026-01-01_09-41-38/types/api.ts
export interface StripeError extends Error {
  type: string;
  code?: string;
  statusCode?: number;
  raw?: unknown;
}
// ...
export interface StripeErrorData extends Error {
  code?: string;
  statusCode?: number;
  requestId?: string;
  type?: string;
  raw?: {
    message: string;
    type: string;
    code?: string;
  };
}